requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
cfgv==3.4.0
charset-normalizer==3.2.0
cloudinary==1.34.0
coverage==7.3.1
cryptography==41.0.3
distlib==0.3.7
dj-database-url==2.1.0
Django==4.2.5
django-cloudinary-storage==0.3.0
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-filter==23.2
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.4
drf-spectacular-sidecar==2023.9.1
factory-boy==3.3.0
Faker==19.6.1
filelock==3.12.4
flake8==6.1.0
flake8-isort==6.0.0
gunicorn==21.2.0
identify==2.5.28
idna==3.4
inflection==0.5.1
isort==5.12.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
mccabe==0.7.0
nodeenv==1.8.0
packaging==23.1
Pillow==10.0.0
platformdirs==3.10.0
pre-commit==3.4.0
psycopg2-binary==2.9.7
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
virtualenv==20.24.5
whitenoise==6.5.0
